% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_Step1_Loading_Data.R
\name{anndata2rds}
\alias{anndata2rds}
\title{Convert anndata to rds}
\usage{
anndata2rds(h5ad.path, metadata.path, spot_diameter = 80, assay = "Spatial")
}
\description{
Convert anndata to rds
}
\details{
https://github.com/STOmics/stereopy/blob/dev/docs/source/_static/annh5ad2rds.R
}
