% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step10_NicheAnalysis.R
\name{CalculateWasserstein}
\alias{CalculateWasserstein}
\title{Calculate the wasserstein distance}
\usage{
CalculateWasserstein(a, b, p = 2)
}
\arguments{
\item{p}{}
}
\description{
Calculate the wasserstein distance
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
