% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step4_cellAnnotation.R
\name{cell_annotation_abc}
\alias{cell_annotation_abc}
\title{cell_annotation_abc}
\usage{
cell_annotation_abc(
  object,
  species = "mmu",
  host = "https://dec2021.archive.ensembl.org/",
  assay = "RNA"
)
}
\arguments{
\item{object}{A Seurat object containing count matrix}

\item{species}{A character indicating the species of the sample}

\item{host}{A character indicating the host to connect to in `useMart` function of
`biomaRt`}

\item{assay}{A character indicating the source `assay` of count matrix}
}
\description{
This function annotates cells using `abcCellmap`
}
