% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step7_assign_cell_cycle.R
\name{cellCycle}
\alias{cellCycle}
\title{Assign Cell Cycle}
\usage{
cellCycle(
  sc_object = NULL,
  counts_matrix = NULL,
  data_matrix = NULL,
  cellcycleCutoff = NULL,
  cellTypeOrders = NULL,
  output.dir = NULL,
  Org = NULL
)
}
\arguments{
\item{sc_object}{A Seurat object containing single-cell RNA sequencing data.}

\item{counts_matrix}{The 'counts' slot in the Seurat object.}

\item{data_matrix}{The 'data' slot in the Seurat object.}

\item{cellcycleCutoff}{The cutoff value for distinguishing between cycling and quiescent cells. Cells with a G1G2Score below this cutoff are considered quiescent.}

\item{cellTypeOrders}{The order of cell types for visualization. If not provided, the function will use the unique cell types in the input Seurat object.}

\item{output.dir}{The path to the directory where the resulting plots will be stored.}

\item{Org}{A character vector specifying the species of cell cycle genes, can be 'mmu' (mouse) or 'hsa' (human).}
}
\value{
Returns an updated Seurat object that includes the assigned cell cycle phases.
}
\description{
This function is used to assign cell cycle phases based on input single-cell RNA sequencing data.
}
\details{
This function assigns cell cycle phases based on the input single-cell RNA sequencing data by analyzing cell cycle-related genes and generates plots of the cell cycle analysis results. 
It uses external data, including lists of cell cycle genes and cell cycle marker genes, which should be prepared and loaded before using the function. 
Note that scran was refered to https://bioconductor.org/packages/release/bioc/vignettes/scran/inst/doc/scran.html. Cell cycle genes are from Dysfunctional CD8 T Cells Form a Proliferative, 
Dynamically Regulated Compartment within Human Melanoma(Table S6)first sheet, supplementary table 1, https://www.nature.com/articles/nature20123#supplementary-information
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
