% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step10_NicheAnalysis.R
\name{Co_occurenceScore}
\alias{Co_occurenceScore}
\title{Co occurence analysis}
\usage{
Co_occurenceScore(
  st_obj,
  features,
  save_path = NULL,
  slice = "slice1",
  method = "Wasserstein",
  p = 2
)
}
\arguments{
\item{st_obj}{SeuratObject}

\item{features}{A vector of features}

\item{save_path}{Save path}

\item{slice}{slice}

\item{method}{correlation or Wasserstein}

\item{p}{The power}
}
\description{
Co occurence analysis
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
