% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step9_violin_plot_for_marker_genes.R
\name{combinedViolinPlot}
\alias{combinedViolinPlot}
\title{Create Violin Plots for Marker Genes}
\usage{
combinedViolinPlot(
  dataMatrix = NULL,
  features = NULL,
  CellTypes = NULL,
  cellTypeOrders = NULL,
  cellTypeColors = NULL,
  Org = NULL,
  output.dir = NULL,
  databasePath = NULL
)
}
\arguments{
\item{dataMatrix}{A data frame or matrix representing the expression data, where rows are cells and columns are genes.}

\item{features}{A character vector specifying the marker genes to plot in the violin plots.}

\item{CellTypes}{A factor vector containing cell type annotations for each cell.}

\item{cellTypeOrders}{A character vector specifying the order of cell types for plotting. Defaults to unique values in `CellTypes`.}

\item{cellTypeColors}{A character vector specifying the colors to use for cell type groups. Defaults to a color palette.}

\item{Org}{A character specifying the organism ('mmu' for mouse, 'hsa' for human).}

\item{output.dir}{The path to the directory where the resulting violin plots will be saved.}

\item{databasePath}{The path to the database required for the analysis.}
}
\value{
None. This function generates and saves violin plots as PDF files.
}
\description{
This function generates violin plots for marker genes across different cell types using the provided expression matrix and cell type annotations.
}
\details{
This function creates violin plots for marker genes across different cell types. You can provide the expression data in the form of a data frame or matrix. The `features` parameter should contain the names of marker genes to be plotted. The `CellTypes` parameter should contain cell type annotations, and `cellTypeOrders` can be used to specify the order in which cell types are displayed in the plots. You can also customize the colors of cell type groups using `cellTypeColors`. The resulting plots are saved to the specified output directory.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
