% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assistantFunctions.R
\name{convertHumanGene}
\alias{convertHumanGene}
\title{Convert human genes to mouse genes}
\usage{
convertHumanGene(human.genes, host = "https://dec2021.archive.ensembl.org/")
}
\arguments{
\item{human.genes}{A vector of human genes}

\item{host}{Host to connect to in `useMart` function}
}
\description{
Convert human genes to mouse genes
}
