% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assistantFunctions.R
\name{convertMatrixHumanGene}
\alias{convertMatrixHumanGene}
\title{convertMatrixHumanGene}
\usage{
convertMatrixHumanGene(counts, host = "https://dec2021.archive.ensembl.org/")
}
\arguments{
\item{counts}{A matrix whose rownames are human genes}

\item{host}{Host to connect to in `useMart` function}
}
\description{
Convert human genes to mouse genes in a matrix with human genes as rownames
}
