% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assistantFunctions.R
\name{convertMouseGene}
\alias{convertMouseGene}
\title{Convert mouse genes to human genes}
\usage{
convertMouseGene(mouse.genes, host = "https://dec2021.archive.ensembl.org/")
}
\arguments{
\item{mouse.genes}{A vector of mouse genes}

\item{host}{Host to connect to in `useMart` function}
}
\description{
Convert mouse genes to human genes
}
