% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step7_CNV_analysis.R
\name{copykatPlot}
\alias{copykatPlot}
\title{copykat plot}
\usage{
copykatPlot(copykat.test, save_path)
}
\arguments{
\item{copykat.test}{The results of `copykat::copykat`}

\item{save_path}{The path to save figures}
}
\description{
copykat plot
}
