% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_doublerFinderwraper.R
\name{doublerFinderwraper}
\alias{doublerFinderwraper}
\title{This function executes the workflow of doubletFinder.}
\usage{
doublerFinderwraper(
  seuratObject = NULL,
  percentage = NULL,
  doublerFinderwraper.PCs = NULL,
  doublerFinderwraper.pN = NULL,
  doublerFinderwraper.pK = NULL
)
}
\arguments{
\item{seuratObject}{The input Seurat object.}

\item{percentage}{Assuming 'percentage' doublet formation rate - tailor for your dataset. The default value is 0.05.}

\item{doublerFinderwraper.PCs}{Which dimensions to use as input features for doubletFinder.}

\item{doublerFinderwraper.pN}{The percentage of real-artifical data for doubletFinder.}

\item{doublerFinderwraper.pK}{The pK parameter controls the doublet cell detection by determining the number of nearest neighbors and influencing the calculation of pANN scores and the final cell classification results. Adjusting the pK value allows optimization of the doublet cell detection process based on specific data and analysis requirements.}
}
\value{
A seurat object with 'singlet' and 'doublet' labels in the column named 'doublet' in meta.data.
}
\description{
Function \code{doublerFinderwraper} takes a Seurat object as input, automatically identifies doublets 
within it, and then adds a column 'doublet' to the meta.data to distinguish between singlets and doublets.
}
\details{
The function \code{doublerFinderwraper} uses doubletFinder to identify doublets within the given Seurat object. It performs the following steps:
1. Identifies pK using paramSweep_v3 and find.pK.
2. Estimates homotypic Doublet Proportion.
3. Runs doubletFinder with varying classification stringencies based on provided parameters.
4. Adds a column 'doublet' to the meta.data to label cells as singlets or doublets.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
