% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_assistantFunctions.R
\name{HemaScopeREnrichment}
\alias{HemaScopeREnrichment}
\title{HemaScopeREnrichment - Function for Gene Ontology (GO) Enrichment Analysis}
\usage{
HemaScopeREnrichment(DEGs = NULL, OrgDb = NULL, output.dir = NULL)
}
\arguments{
\item{DEGs}{A data frame containing differential gene expression information with columns "cluster" and "gene". The "cluster" column specifies the cluster to which each gene belongs, and the "gene" column contains the gene names.}

\item{OrgDb}{A database for gene annotation, used to map gene symbols to GO terms. e.g. It should be 'org.Mm.eg.db' for mmu and 'org.Hs.eg.db' for hsa.}

\item{output.dir}{The directory path where the output files (PDF and CSV) will be saved.}
}
\value{
The function does not return any values directly, but it saves PDF files with barplots showing significant GO enrichment results for each cluster, and CSV files containing the enriched GO terms and associated statistics.
}
\description{
This function performs Gene Ontology (GO) enrichment analysis for a set of differentially expressed genes (DEGs) grouped by clusters. It generates GO enrichment results in terms of Biological Process (BP) category, including barplots and CSV files.
}
\details{
The function takes a data frame of DEGs and performs GO enrichment analysis separately for each cluster defined in the "cluster" column. It utilizes the "enrichGO" function to calculate GO enrichment, and generates barplots of significant enriched GO terms for each cluster. The results are saved as both PDF files and CSV files, providing insights into the biological processes associated with the DEGs in different clusters.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
