% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step8_heterogeneity.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Measure Cell Heterogeneity Using Spearman Correlation}
\usage{
heterogeneity(
  expression_matrix = NULL,
  cell_types_groups = NULL,
  cellTypeOrders = NULL,
  output.dir = NULL
)
}
\arguments{
\item{expression_matrix}{A numeric matrix representing the expression data, where rows are genes and columns are cells.}

\item{cell_types_groups}{A data frame specifying cell type annotations for each cell, including cell type labels and group information.}

\item{cellTypeOrders}{The order of cell types for visualization. If not provided, the function will use the unique cell types in the input cell_types_groups.}

\item{output.dir}{The path to the directory where the resulting plots will be saved.}
}
\value{
None. This function generates and saves plots to the specified output directory.
}
\description{
This function calculates the cell heterogeneity based on Spearman correlation using the input expression matrix and cell type group annotations.
}
\details{
This function quantifies cell heterogeneity by computing Spearman correlation coefficients between cells within the same cell type groups. It generates plots visualizing the heterogeneity measurements for different cell types and groups. The `expression_matrix` should be appropriately preprocessed and filtered before using this function.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
