% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{init_miniconda}
\alias{init_miniconda}
\title{Init miniconda for scRNA-seq pipeline and basic ST pipline}
\usage{
init_miniconda()
}
\description{
Init miniconda for scRNA-seq pipeline and basic ST pipline
}
