% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step10_calculate_lineage_scores.R
\name{lineageScores}
\alias{lineageScores}
\title{Calculate Lineage Scores}
\usage{
lineageScores(
  expression_matrix = NULL,
  cellTypes = NULL,
  cellTypes_orders = NULL,
  cellTypes_colors = NULL,
  groups = NULL,
  groups_orders = NULL,
  groups_colors = NULL,
  lineage.genelist = NULL,
  lineage.names = NULL,
  Org = NULL,
  output.dir = NULL,
  databasePath = NULL
)
}
\arguments{
\item{expression_matrix}{A data frame or matrix representing the expression data, where rows are cells and columns are genes.}

\item{cellTypes}{A character vector specifying cell type annotations for each cell. e.g. c("HSC","HSC","HSC","MPP1","MPP2","MPP2","MPP2","MPP3","MPP4","MEP","GMP", "MKP","CMP","CLP","LSKlow", ...)}

\item{cellTypes_orders}{A character vector specifying the order of cell types for plotting. e.g. c("HSC","MPP1","MPP2","MPP3","MPP4","MEP","GMP", "MKP","CMP","CLP","LSKlow")}

\item{cellTypes_colors}{A character vector specifying the colors to use for cell type groups. e.g. c("HSC" = '#006d2c',"MPP1" = '#4292c6',"MPP2"= '#810f7c',"MPP3" = '#fec44f',"MPP4" = '#dd3497',"MEP" = '#ef6548',"GMP" = '#993404', "MKP" = '#8073ac',"CMP" = '#8c510a',"CLP" = '#b2df8a',"LSKlow" = '#d53e4f').}

\item{groups}{A character vector specifying groups or clusters within each cell type.}

\item{groups_orders}{A character vector specifying the order of groups or clusters for plotting.}

\item{groups_colors}{A character vector specifying the colors to use for group or cluster annotations. e.g. c('group1'='#d73027','group2'='#2171b5')}

\item{lineage.genelist}{A list of gene sets representing lineage markers.}

\item{lineage.names}{A character vector specifying the names of the lineages.}

\item{Org}{A character specifying the organism ('mmu' for mouse, 'hsa' for human).}

\item{output.dir}{The path to the directory where the resulting visualizations and scores will be saved.}

\item{databasePath}{The path to the database required for the analysis.}
}
\value{
None. This function generates and saves visualizations and lineage scores as PDF and CSV files.
}
\description{
This function calculates lineage scores based on provided expression data and cell type annotations. It also generates visualizations of lineage scores and gene expression patterns.
}
\details{
This function calculates lineage scores for specified gene sets based on the provided expression data. It then generates a heatmap of lineage scores and a heatmap of gene expression patterns. You can customize the order and colors of cell types, groups, and clusters for visualization. The resulting visualizations and lineage scores are saved to the specified output directory.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
