% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step4_query_data.R
\name{mapDataToRef}
\alias{mapDataToRef}
\title{Map Data from Query Object to Reference Object}
\usage{
mapDataToRef(
  ref_object = NULL,
  ref_labels = NULL,
  query_object = NULL,
  PCs = NULL,
  output.dir = NULL
)
}
\arguments{
\item{ref_object}{A reference Seurat object containing the reference data.}

\item{ref_labels}{A vector or factor specifying the labels for the reference data.}

\item{query_object}{A query Seurat object containing the data to be mapped to the reference.}

\item{PCs}{A vector specifying the principal components to be used for mapping.}

\item{output.dir}{The directory where the mapping results and plots will be saved.}
}
\value{
mapDataToRef returns the query Seurat object with added prediction scores and IDs.
}
\description{
Function \code{mapDataToRef} performs data mapping from a query Seurat object to a reference Seurat object.
}
\details{
This function maps data from a query Seurat object to a reference Seurat object using the specified principal components (PCs). It identifies transfer anchors between the reference and query objects, transfers data, and adds predictions to the query object. The mapping results are visualized and saved as plots in the output directory. Note that ref_object and query_object should have sample type feature names. Refer to https://satijalab.org/seurat/archive/v3.0/integration.html
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
