% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERFISH_Step1_Loading_Data.R
\name{MERFISH_Loading_Data}
\alias{MERFISH_Loading_Data}
\title{Loading MERFISH data}
\usage{
MERFISH_Loading_Data(
  input.data.dir,
  output.dir = ".",
  tech = "Vizgen",
  fov = "fov",
  rds.file = FALSE,
  assay = NULL,
  Vizgen.z = 3L,
  Akoya.type = "inform"
)
}
\arguments{
\item{output.dir}{A character of path to store the h5ad file}

\item{tech}{'Vizgen', 'Xenium', 'Nonastring', or 'Akoya'}

\item{fov}{A character of the name of the sample}

\item{rds.file}{Whether it is a rds file}

\item{assay}{Name of assay}

\item{Vizgen.z}{Z-index to load; must be between 0 and 6, inclusive}

\item{Akoya.type}{'processor', 'inform', or 'qupath'}

\item{input.data.dirs}{A character of data path}
}
\description{
Loading MERFISH data
}
\details{
This function loads MERFISH data
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
