% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERFISH_Step5_SpatiallyVariableFeatures.R
\name{MERFISH_SpatiallyVariableFeatures}
\alias{MERFISH_SpatiallyVariableFeatures}
\title{Find spatially variable features}
\usage{
MERFISH_SpatiallyVariableFeatures(
  st_obj,
  output.dir = ".",
  assay = "SCT",
  selection.method = "moransi",
  n.top.show = 10,
  n.col = 5,
  verbose = FALSE
)
}
\arguments{
\item{st_obj}{The SeuratObject of MERFISH data}

\item{output.dir}{A character of path to store the results and figures}

\item{assay}{The assay used}

\item{selection.method}{Method for selecting spatially variable features,
`markvariogram` or `moransi`}

\item{n.top.show}{The number of top genes shown in the figure}

\item{n.col}{The number of columns in the output figure}

\item{verbose}{verbose as `Seurat`}
}
\description{
Find spatially variable features
}
\details{
This function finds spatially variable features and plots the distribution of
these genes
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
