% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assistantFunctions.R
\name{mySpatialFeaturePlot}
\alias{mySpatialFeaturePlot}
\title{Plot features on the image}
\usage{
mySpatialFeaturePlot(
  st_obj,
  features,
  legend.name = NULL,
  legend.color = colorRampPalette(colors = rev(x = brewer.pal(n = 11, name = "RdYlBu"))),
  n.col = 4,
  ...
)
}
\description{
Plot features on the image
}
