% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step5_SpatiallyVariableFeatures.R
\name{mySpatiallyVariableFeatures}
\alias{mySpatiallyVariableFeatures}
\title{SpatiallyVariableFeatures revision}
\usage{
mySpatiallyVariableFeatures(
  object,
  assay = "SCT",
  selection.method = "moransi"
)
}
\description{
SpatiallyVariableFeatures revision
}
