% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step10_NicheAnalysis.R
\name{NicheCluster}
\alias{NicheCluster}
\title{Cluster niches}
\usage{
NicheCluster(st_obj, features, save_path = NULL, kmeans.n = 4)
}
\arguments{
\item{st_obj}{SeuratObject}

\item{features}{A vector of features}

\item{save_path}{Save path}

\item{kmeans.n}{The number of clusters}
}
\description{
Cluster niches
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
