% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenXGR_oColormap.R
\name{oColormap}
\alias{oColormap}
\title{Function to define a colormap}
\usage{
oColormap(
  colormap = c("bwr", "jet", "gbr", "wyr", "br", "yr", "rainbow", "wb", "heat",
    "terrain", "topo", "cm", "ggplot2", "jet.top", "jet.bottom", "jet.both", "spectral",
    "ggplot2.top", "ggplot2.bottom", "ggplot2.both", "RdYlBu", "brewer.BrBG",
    "brewer.PiYG", "brewer.PRGn", "brewer.PuOr", "brewer.RdBu", "brewer.RdGy",
    "brewer.RdYlBu", "brewer.RdYlGn", "brewer.Spectral", "brewer.Blues", "brewer.BuGn",
    "brewer.BuPu", "brewer.GnBu", "brewer.Greens", "brewer.Greys", "brewer.Oranges",
    "brewer.OrRd", "brewer.PuBu", "brewer.PuBuGn", 
     "brewer.PuRd", "brewer.Purples",
    "brewer.RdPu", "brewer.Reds", "brewer.YlGn", "brewer.YlGnBu", "brewer.YlOrBr",
    "brewer.YlOrRd", "rainbow_hcl", "heat_hcl", "terrain_hcl", "diverge_hcl", "hcl_br",
    "hcl_bp", "hcl_bb", "hcl_gp", "hcl_go", "hcl_cp", "hcl_cy", "hcl_co", "sci_jco",
    "sci_lancet", "sci_nejm", "sci_locuszoom"),
  interpolate = c("spline", "linear"),
  data = NULL,
  zlim = NULL
)
}
\arguments{
\item{colormap}{short name for the colormap. It can be one of "jet" (jet colormap), "bwr" (blue-white-red colormap), "gbr" (green-black-red colormap), "wyr" (white-yellow-red colormap), "br" (black-red colormap), "yr" (yellow-red colormap), "wb" (white-black colormap), "rainbow" (rainbow colormap, that is, red-yellow-green-cyan-blue-magenta), and "ggplot2" (emulating ggplot2 default color palette). Alternatively, any hyphen-separated HTML color names, e.g. "lightyellow-orange" (by default), "blue-black-yellow", "royalblue-white-sandybrown", "darkgreen-white-darkviolet". A list of standard color names can be found in \url{https://html-color-codes.info/color-names/index.html}. It can also be a function of 'grDevices::colorRampPalette'. It can also be "brewer.*" (RColorBrewer palette; see RColorBrewer::display.brewer.all()). It can be colorspace defaults ("rainbow_hcl","heat_hcl","terrain_hcl","diverge_hcl") or other useful ones ("hcl_br","hcl_bp","hcl_bb","hcl_gp","hcl_go","hcl_cp","hcl_cy","hcl_co")}

\item{interpolate}{use spline or linear interpolation}

\item{data}{NULL or a numeric vector}

\item{zlim}{the minimum and maximum z/patttern values for which colors should be plotted, defaulting to the range of the finite values of z. Each of the given colors will be used to color an equispaced interval of this range. The midpoints of the intervals cover the range, so that values just outside the range will be plotted}
}
\value{
palette.name (a function that takes an integer argument for generating that number of colors interpolating the given sequence) or mapped colors if data is provided.
}
\description{
\code{oColormap} is supposed to define a colormap. It returns a function, which will take an integer argument specifying how many colors interpolate the given colormap.
}
\note{
The input colormap includes: 
\itemize{
\item{"jet": jet colormap}
\item{"bwr": blue-white-red}
\item{"gbr": green-black-red}
\item{"wyr": white-yellow-red}
\item{"br": black-red}
\item{"yr": yellow-red}
\item{"wb": white-black}
\item{"rainbow": rainbow colormap, that is, red-yellow-green-cyan-blue-magenta}
\item{"ggplot2": emulating ggplot2 default color palette}
\item{"spectral": emulating RColorBrewer spectral color palette}
\item{Alternatively, any hyphen-separated HTML color names, e.g. "blue-black-yellow", "royalblue-white-sandybrown", "darkblue-lightblue-lightyellow-darkorange", "darkgreen-white-darkviolet", "darkgreen-lightgreen-lightpink-darkred". A list of standard color names can be found in \url{https://html-color-codes.info/color-names/index.html}}
}
}
\examples{
# 1) define "blue-white-red" colormap
palette.name <- oColormap(colormap="bwr")
# use the return function "palette.name" to generate 10 colors spanning "bwr"
palette.name(10)

# 2) define default colormap from ggplot2
palette.name <- oColormap(colormap="ggplot2")
# use the return function "palette.name" to generate 3 default colors used by ggplot2
palette.name(3)

# 3) define brewer colormap called "RdYlBu"
palette.name <- oColormap(colormap="RdYlBu")
# use the return function "palette.name" to generate 3 default colors used by ggplot2
palette.name(3)

# 4) return mapped colors
oColormap(colormap="RdYlBu", data=runif(5))
}
\seealso{
\code{\link{oColormap}}
}
