% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenXGR_oIG2TB.R
\name{oIG2TB}
\alias{oIG2TB}
\title{Function to convert an igraph into a tibble for nodes or edges}
\usage{
oIG2TB(ig, what = c("edges", "nodes"))
}
\arguments{
\item{ig}{an "igraph" object}

\item{what}{what to extract. It can be "edges" for edges and "nodes" for nodes}
}
\value{
a tibble object
}
\description{
\code{oIG2TB} is supposed to convert an igraph into a tibble for nodes or edges.
}
\note{
none
}
\examples{
set.seed(825)
ig <- sample_pa(20)
V(ig)$name <- seq(1,vcount(ig))
ig \%>\% oIG2TB('edges')
ig \%>\% oIG2TB('nodes')
}
\seealso{
\code{\link{oIG2TB}}
}
