% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenXGR_oNetInduce.R
\name{oNetInduce}
\alias{oNetInduce}
\title{Function to generate a subgraph induced by given vertices and their k nearest neighbors}
\usage{
oNetInduce(
  ig,
  nodes_query,
  knn = 0,
  remove.loops = FALSE,
  largest.comp = TRUE,
  min.comp.size = 1,
  verbose = TRUE
)
}
\arguments{
\item{ig}{an object of class "igraph"}

\item{nodes_query}{the vertices for which the calculation is performed}

\item{knn}{an integeter specifying how many k steps are used to find the nearest neighbours of the given vertices. By default, knn is set to zero; it means no neighbors will be considered. When knn is 1, the immediate neighbors of the given vertices will be also considered for inducing the subgraph. The same is true when knn is 2, etc}

\item{remove.loops}{logical to indicate whether the loop edges are to be removed. By default, it sets to false}

\item{largest.comp}{logical to indicate whether the largest component is only retained. By default, it sets to true for the largest component being left}

\item{min.comp.size}{an integer specifying the minimum size of component that will be retained. This parameter only works when setting the false to keep the largest component. By default, it sets to 1 meaning all nodes will be retained}

\item{verbose}{logical to indicate whether the messages will be displayed in the screen. By default, it sets to true for display}
}
\value{
\itemize{
 \item{\code{subg}: an induced subgraph, an object of class "igraph". Appended with a node attribute 'comp' if multiple components are kept}
}
}
\description{
\code{oNetInduce} is supposed to produce a subgraph induced by given vertices and its k nearest neighbors. The input is a graph of "igraph" or "graphNET" object, a list of the vertices of the graph, and a k value for finding k nearest neighbors for these vertices. The output is a subgraph induced by given vertices plus their k neighbours. The resultant subgraph inherits the class from the input one. The induced subgraph contains exactly the vertices of interest, and all the edges between them.
}
\note{
The given vertices plus their k nearest neighbors will be used to induce the subgraph.
}
\examples{
\dontrun{
# 1) generate a random graph according to the ER model
g <- erdos.renyi.game(100, 1/100)
V(g)$name <- as.character(V(g))

# 2) select the first 10 vertices as the query nodes
nodes_query <- as.character(1:4)

# 3) produce the induced subgraph only based on the nodes in query
subg <- oNetInduce(g, nodes_query, knn=0)

# 4) produce the induced subgraph based on the nodes in query are their immediate neighbours
subg <- oNetInduce(g, nodes_query, knn=1)
}
}
\seealso{
\code{\link{oNetInduce}}
}
