% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenXGR_oRDS.R
\name{oRDS}
\alias{oRDS}
\title{Function to read RDS files}
\usage{
oRDS(RDS = NULL, verbose = TRUE, placeholder = NULL, guid = NULL)
}
\arguments{
\item{RDS}{which RDS to load. To support the remote reading of a compressed RDS file, it must be compressed via the gzip method}

\item{verbose}{logical to indicate whether the messages will be displayed in the screen. By default, it sets to TRUE for display}

\item{placeholder}{the characters to tell the placeholder of RDS files}

\item{guid}{a valid (5-character) Global Unique IDentifier for an OSF project. For example, 'gskpn' (see 'https://osf.io/gskpn'). If a valid provided and the query matched, it has priority over the one specified via placeholder}
}
\value{
the loaded RDS. If the data cannot be loaded, it returns NULL.
}
\description{
\code{oRDS} is supposed to read RDS files.
}
\note{
To enable 'guid', please also install a package "osfr" via \code{BiocManager::install("osfr",dependencies=TRUE)}.
}
\examples{
placeholder <- 'http://www.comptransmed.pro/bigdata_fdb'
placeholder <- 'http://www.comptransmed.pro/bigdata_ctm'
placeholder <- 'http://www.comptransmed.pro/bigdata_pich'
placeholder <- 'http://www.comptransmed.pro/bigdata_pia'
\dontrun{
org.Hs.eg <- oRDS('org.Hs.eg', placeholder=placeholder)
}
}
\seealso{
\code{\link{oRDS}}
}
