% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step6_Interaction.R
\name{plotStrengthAndPvalue_total}
\alias{plotStrengthAndPvalue_total}
\title{Plot total strength and P-value}
\usage{
plotStrengthAndPvalue_total(
  commot_matrix.list,
  commot_pvalue.list,
  save_path = NULL,
  figure_name = "Dotplot",
  between.cluster = TRUE,
  height = 4,
  width = 4
)
}
\arguments{
\item{save_path}{The path to save outputs}

\item{figure_name}{The name of figures}

\item{commot_matrix}{The output file of `Commot`}

\item{commot_pvalue}{The output file of `Commot`}
}
\description{
Plot total strength and P-value
}
