% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step6_Interaction.R
\name{plotStrengthAndPvalue_within}
\alias{plotStrengthAndPvalue_within}
\title{Plot the strength and P-value}
\usage{
plotStrengthAndPvalue_within(
  commot_matrix,
  commot_pvalue,
  save_path = NULL,
  figure_name = "Dotplot",
  height = 4,
  width = 4
)
}
\arguments{
\item{commot_matrix}{The output file of `Commot`}

\item{commot_pvalue}{The output file of `Commot`}

\item{save_path}{The path to save outputs}

\item{figure_name}{The name of figures}
}
\description{
Plot the strength and P-value
}
