% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{python.path.sc}
\alias{python.path.sc}
\title{The path to the miniconda environment of scRNA-seq pipeline}
\usage{
python.path.sc()
}
\description{
The path to the miniconda environment of scRNA-seq pipeline
}
