% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python.R
\name{python.path.stereo}
\alias{python.path.stereo}
\title{The path to the miniconda environment of stereo-seq pipeline}
\usage{
python.path.stereo()
}
\description{
The path to the miniconda environment of stereo-seq pipeline
}
