% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MERFISH_Step2_QC.R
\name{QC_MERFISH}
\alias{QC_MERFISH}
\title{Quality control on ST data}
\usage{
QC_MERFISH(
  st_obj,
  output.dir = ".",
  min.gene = 20,
  min.nUMI = 50,
  max.gene = Inf,
  max.nUMI = Inf,
  min.spot = 3,
  species = "human",
  bool.remove.mito = TRUE,
  SpatialColors = colorRampPalette(colors = brewer.pal(n = 9, name = "YlOrRd"))
)
}
\arguments{
\item{st_obj}{The SeuratObject of merfish data}

\item{output.dir}{A character of path to store the results and figures}

\item{min.gene}{An integer representing the minimum number of genes detected in a spot}

\item{min.nUMI}{An integer representing the minimum number of nUMI detected in a spot}

\item{max.gene}{An integer representing the maximum number of genes detected in a spot}

\item{max.nUMI}{An integer representing the maximum number of nUMI detected in a spot}

\item{min.spot}{An integer representing the minimum number of spots expressing each gene}

\item{species}{A character representing the species of sample, `human` or `mouse`}

\item{bool.remove.mito}{A bool value indicating whether removing mitochondrial gene}

\item{SpatialColors}{A function that interpolates a set of given colors to create new color palettes and color ramps}
}
\description{
Quality control on ST data
}
\details{
This function performs basic quality control on the SeuratObject of merfish data and
returns the filtered SeuratObject
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
