% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step2_QC_multiple_scRNASeq.R
\name{QC_multiple_scRNASeq}
\alias{QC_multiple_scRNASeq}
\title{The quality control step for multiple scRNA-Seq datasets.}
\usage{
QC_multiple_scRNASeq(
  seuratObjects = NULL,
  datasetID = NULL,
  output.dir = NULL,
  Step2_Quality_Control.RemoveBatches = FALSE,
  Step2_Quality_Control.RemoveDoublets = FALSE,
  nFeature_RNA.limit = NULL,
  percent.mt.limit = NULL,
  scale.factor = NULL,
  nfeatures = NULL,
  ndims = NULL,
  vars.to.regress = NULL,
  PCs = NULL,
  resolution = NULL,
  n.neighbors = NULL,
  percentage = NULL,
  doublerFinderwraper.PCs = NULL,
  doublerFinderwraper.pN = NULL,
  doublerFinderwraper.pK = NULL
)
}
\arguments{
\item{seuratObjects}{The input Seurat objects.}

\item{datasetID}{The ID of each dataset (e.g. c('ctrl', 'case')). Please do not include ':' in the datasetID.}

\item{output.dir}{The path of directory for saving putputs.}

\item{Step2_Quality_Control.RemoveBatches}{A logical value indicating whether to perform batch removal. Default is FALSE.}

\item{Step2_Quality_Control.RemoveDoublets}{A logical value indicating whether to remove doublets. Default is FALSE.}

\item{nFeature_RNA.limit}{The cutoff of the minimum number of detected genes in each cell.}

\item{percent.mt.limit}{The cutoff of the maximum percentage of mitochondria genes in each cell.}

\item{scale.factor}{The scale factor for the 'data' slot in the seurat object.}

\item{nfeatures}{The number of selected highly variable features for down stream analysis.}

\item{ndims}{The number of principle components in PCA.}

\item{vars.to.regress}{Variables to regress out (previously latent.vars in RegressOut). For example, nUMI, or percent.mito. (ScaleData in Seurat)}

\item{PCs}{Which dimensions to use as input features.(RunTSNE and RunUMAP in Seurat)}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities. (FindClusters in Seurat)}

\item{n.neighbors}{Defines k for the k-nearest neighbor algorithm. (FindNeighbors in Seurat)}

\item{percentage}{Assuming 'percentage' doublet formation rate - tailor for your dataset. The default value is 0.05.}

\item{doublerFinderwraper.PCs}{Which dimensions to use as input features for doubletFinder.}

\item{doublerFinderwraper.pN}{The percentage of real-artifical data for doubletFinder.}

\item{doublerFinderwraper.pK}{The pK parameter controls the doublet cell detection by determining the number of nearest neighbors and influencing the calculation of pANN scores and the final cell classification results. Adjusting the pK value allows optimization of the doublet cell detection process based on specific data and analysis requirements.}
}
\value{
Return a Seurat object containing data after quality control.
}
\description{
Function \code{QC_multiple_scRNASeq} perform the quality control step for multiple scRNA-Seq datasets,
which includes normalizing data, finding variable features, scaling data, doing principle component analysis,
calculating K nearest neighbors, clustering, removing doublets, and integrating multiple datasets.
}
\details{
The function \code{QC_multiple_scRNASeq} takes Seurat objects representing scRNA-Seq datasets as input and performs the following quality control steps:
1. Normalize data using the LogNormalize method.
2. Find variable features using the vst method.
3. Scale data using the identified variable features and specified variables to regress out.
4. Perform principal component analysis (PCA) on the scaled data.
5. Find K nearest neighbors based on PCA dimensions.
6. Perform clustering analysis based on the found neighbors.
7. Optionally, remove doublets using doubletFinder.
8. Optionally, integrate multiple datasets by removing batch effects.

The resulting Seurat object contains data after quality control.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
