% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_assistantFunctions.R
\name{Rds2H5}
\alias{Rds2H5}
\title{Save object to h5 file and spatial file}
\usage{
Rds2H5(object, file.dir = ".", assay = "Spatial", slice = "slice1")
}
\arguments{
\item{object}{A Seuratobject}

\item{file.dir}{A path to store outputs}

\item{assay}{The assay to store}

\item{slice}{The slice to store}
}
\description{
Save object to h5 file and spatial file
}
