% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step4_cellAnnotation.R
\name{run_cell_annotation}
\alias{run_cell_annotation}
\title{run_cell_annotation}
\usage{
run_cell_annotation(
  object,
  split = TRUE,
  frag_num = 5000,
  species = "mmu",
  host = "https://dec2021.archive.ensembl.org/",
  assay = "RNA",
  output.dir = NULL
)
}
\arguments{
\item{object}{A Seurat object containing count matrix}

\item{split}{A bool value indicating whether to predict cells in batches}

\item{frag_num}{A integer indicating the number of cells in one batch}

\item{species}{A character indicating the species of the sample}

\item{host}{A character indicating the host to connect to in `useMart` function of `biomaRt`}

\item{assay}{A character indicating the source `assay` of count matrix}
}
\description{
This function annotates cells using [abcCellmap](http://scrna.sklehabc.com/) and
returns the Seurat object with annotation labels. To save memory, the function
can perform batch predictions on the data
}
\details{
The following documents are from 'abcCellmap' package: Users can predict the cell types of hematopoietic cells by implementing two approaches (Scmap and Seurat). 
Cells in our ABC are labeled by 43 different RNA clusters according to unsupervised clustering of single-cell transcriptional profiles, 
and also labeled by 32 immunophenotypic cell types, involving HSPC, B cell, T cell, NK cell, Neutrophil, Monocyte and Erythrocyte population. The format of result are as follows:
queryCell, Seurat.RNACluster, Seurat.RNACluster.score, Seurat.Immunophenotype, Seurat.Immunophenotype.score, 
scmap.RNACluster, scmap.RNACluster.score, scmap.Immunophenotype, scmap.Immunophenotype.score, scmap.Cell, scmap.Cell.score, pertype.
Of which, "queryCell" is the cell information in the query data,
"Seurat.RNACluster" is the RNA cluster predicted by Seurat,
"Seurat.RNACluster.score" is the prediction score of RNA cluster by Seurat,
"Seurat.Immunophenotype" is the immunophenotypic cell type predicted by Seurat,
"Seurat.Immunophenotype.score" is the prediction score of immunophenotypic cell type by Seurat,
"scmap.RNACluster" is the RNA cluster predicted by scmap,
"scmap.RNACluster.score" is the prediction score of RNA cluster by scmap,
"scmap.Immunophenotype" is the immunophenotypic cell type predicted by scmap,
"scmap.Immunophenotype.score" is the prediction score of immunophenotypic cell type by scmap,
"scmap.Cell" is the nearest single cell in our ABC reference predicted by scmap,
"scmap.Cell.score" is the prediction score of the nearest cell by scmap,
"pertype" means the percentages of top 2 immunophenotypic cell types in corresponding Seurat.RNACluster result.
}
