% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step14_run_CellChat.R
\name{run_CellChat}
\alias{run_CellChat}
\title{run_CellChat - Function for Cell-Cell Communication Analysis with CellChat}
\usage{
run_CellChat(
  data.input = NULL,
  labels = NULL,
  cell.orders = NULL,
  cell.colors = NULL,
  sample.names = NULL,
  Org = NULL,
  sorting = FALSE,
  output.dir = NULL
)
}
\arguments{
\item{data.input}{A matrix of expression data, where rows represent genes and columns represent cells. Row names should be in the format of gene symbols.}

\item{labels}{A vector of cluster labels for each cell, corresponding to the columns of data.input.}

\item{cell.orders}{A character vector specifying the order of cell types or clusters in the analysis.}

\item{cell.colors}{A character vector specifying colors for cell types or clusters in the analysis.}

\item{sample.names}{A vector of sample or cell names, corresponding to the columns of data.input.}

\item{Org}{A string indicating the organism used in the analysis. It should be either "mmu" (mouse) or "hsa" (human).}

\item{sorting}{A logical value indicating whether to consider cell population size in communication analysis.}

\item{output.dir}{The directory path where the output files, including plots and data, will be saved.}
}
\value{
The function does not return any values directly, but it saves PDF files with various network visualizations, plots, and data files that can be used for further analysis.
}
\description{
This function performs cell-cell communication analysis using the CellChat package. It takes expression data and cluster labels as input, identifies cell-cell communication networks, and visualizes the results, including interaction networks and signaling pathways.
}
\details{
The function takes expression data, cluster labels, and other parameters to perform cell-cell communication analysis using the CellChat package. It includes the following steps:

1. Data input and preprocessing.
2. Initialization of a CellChat object.
3. Set the ligand-receptor interaction database based on the specified organism.
4. Preprocess the expression data for cell-cell communication analysis.
5. Identify overexpressed genes and interactions.
6. Project data based on protein-protein interaction networks.
7. Inference of cell-cell communication network.
8. Visualization of the communication network.
9. Systems analysis of cell-cell communication network.

The results are saved as PDF files containing various network visualizations and analysis plots.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
