% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step5_run_HematoMap.R
\name{run_HematoMap}
\alias{run_HematoMap}
\title{Dimension reduction by HematoMap.}
\usage{
run_HematoMap(sc_object = NULL, output.dir = NULL)
}
\arguments{
\item{sc_object}{A Seurat object containing single-cell RNA-seq data.}

\item{output.dir}{The directory where cluster tree plots will be saved.}
}
\value{
run_HematoMap returns cluster tree plots in both PDF and PNG formats.
}
\description{
Function \code{run_HematoMap} performs dimension reduction using HematoMap.
}
\details{
This function performs dimension reduction using HematoMap on a Seurat object.
It generates cluster tree plots and saves them as PDF files and PNG files in the specified output directory.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
