% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step12_run_monocle.R
\name{run_monocle}
\alias{run_monocle}
\title{Run Monocle Analysis}
\usage{
run_monocle(
  cellData = NULL,
  phenoData = NULL,
  featureData = NULL,
  lowerDetectionLimit = 0.5,
  expressionFamily = VGAM::negbinomial.size(),
  cellTypes = NULL,
  monocle.orders = NULL,
  monocle.colors = NULL,
  output.dir = getwd()
)
}
\arguments{
\item{cellData}{A matrix of gene expression values, where columns represent cells and rows represent genes.}

\item{phenoData}{A data frame containing cell metadata, such as cell labels or other relevant information.}

\item{featureData}{A data frame containing information about features (genes) in the dataset.}

\item{lowerDetectionLimit}{The lower detection limit for gene expression. Genes with expression values below this limit will be treated as non-detected.}

\item{expressionFamily}{The family of the expression distribution used in Monocle analysis.}

\item{cellTypes}{A character vector specifying cell types or labels used for coloring in trajectory plots.}

\item{monocle.orders}{A character vector specifying the order of cell types in the Monocle analysis.}

\item{monocle.colors}{A character vector specifying colors for cell types in trajectory plots.}

\item{output.dir}{The directory where the Monocle analysis results and plots will be saved.}
}
\value{
None. This function generates and saves Monocle analysis results and trajectory plots.
}
\description{
Function \code{run_monocle} runs the Monocle analysis on single-cell RNA-seq data to infer cell trajectories and explore gene expression changes along these trajectories.
}
\details{
This function performs Monocle analysis on single-cell RNA-seq data. It involves several steps:

1. Creating a Monocle cellDataSet using the provided cellData, phenoData, and featureData.
2. Estimating size factors, dispersions, and detecting highly variable genes.
3. Performing differential gene expression analysis to identify genes associated with cell state changes.
4. Ordering cells along the inferred trajectories and reducing dimensionality.
5. Generating and saving trajectory plots, including cell trajectory by "State" and by "Cell Types."
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
