% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step5_run_phateR.R
\name{run_phateR}
\alias{run_phateR}
\title{Dimension reduction by phateR.}
\usage{
run_phateR(
  sc_object = NULL,
  output.dir = NULL,
  pythonPath = NULL,
  phate.knn = 50,
  phate.npca = 20,
  phate.t = 10,
  phate.ndim = 2
)
}
\arguments{
\item{sc_object}{A Seurat object containing single-cell RNA-seq data.}

\item{output.dir}{The directory where the dimension reduction plot will be saved.}

\item{pythonPath}{The path to the Python executable if required for running phateR.}

\item{phate.knn}{The number of nearest neighbors to consider in the phateR algorithm. Default 50.}

\item{phate.npca}{The number of principal components to use in the phateR algorithm. Default 20.}

\item{phate.t}{The t-value for the phateR algorithm, which controls the level of exploration. Default 10.}

\item{phate.ndim}{The number of dimensions for the output embedding in the phateR algorithm. Default 2.}
}
\value{
run_phateR returns the dimension reduction plot in both PDF and PNG formats.
}
\description{
Function \code{run_phateR} performs dimension reduction using phateR algorithm.
}
\details{
This function performs dimension reduction using the phateR algorithm on a Seurat object.
It generates a dimension reduction plot and saves it as a PDF and PNG file in the specified output directory.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
