% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step13_run_SCENIC.R
\name{run_SCENIC}
\alias{run_SCENIC}
\title{Run SCENIC Analysis}
\usage{
run_SCENIC(
  countMatrix = NULL,
  cellTypes = NULL,
  datasetID = NULL,
  cellTypes_colors = NULL,
  cellTypes_orders = NULL,
  groups_colors = NULL,
  groups_orders = NULL,
  Org = NULL,
  output.dir = NULL,
  pythonPath = NULL,
  databasePath = NULL
)
}
\arguments{
\item{countMatrix}{A matrix containing the raw counts of the single-cell RNA-seq data.}

\item{cellTypes}{A character vector specifying the cell types or labels for each cell.}

\item{datasetID}{A character vector specifying the dataset IDs for each cell.}

\item{cellTypes_colors}{A named vector of colors for cell type visualization.}

\item{cellTypes_orders}{A character vector specifying the desired order of cell types.}

\item{groups_colors}{A named vector of colors for grouping visualization.}

\item{groups_orders}{A character vector specifying the desired order of groups.}

\item{Org}{A character vector specifying the organism ('mmu' for mouse or 'hsa' for human).}

\item{output.dir}{The directory where the SCENIC results and output files will be saved.}

\item{pythonPath}{The path to the Python environment to use for the analysis.}
}
\value{
None. This function saves SCENIC results and output files in the specified output directory.
}
\description{
Function \code{run_SCENIC} performs SCENIC (Single-Cell Regulatory Network Inference and Clustering) analysis on a given single-cell RNA-seq dataset.
}
\details{
This function runs SCENIC analysis, including the construction of a co-expression network, gene filtering, correlation, and the GENIE3 algorithm to infer regulatory networks.
It also saves various output files for downstream analysis and visualization.
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
