% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step12_run_slingshot.R
\name{run_slingshot}
\alias{run_slingshot}
\title{Run Slingshot Analysis}
\usage{
run_slingshot(
  slingshot.PCAembeddings = NULL,
  slingshot.cellTypes = NULL,
  slingshot.start.clus = NULL,
  slingshot.end.clus = NULL,
  slingshot.colors = NULL,
  output.dir = getwd()
)
}
\arguments{
\item{slingshot.PCAembeddings}{A matrix containing the PCA embeddings of the single-cell data, typically obtained from dimensionality reduction techniques like PCA.}

\item{slingshot.cellTypes}{A character vector specifying cell types or labels for each cell.}

\item{slingshot.start.clus}{A character vector specifying the initial cluster(s) from which cell trajectories should start.}

\item{slingshot.end.clus}{A character vector specifying the target cluster(s) where cell trajectories should end.}

\item{slingshot.colors}{A vector of colors corresponding to cell types for plotting. If not provided, default colors will be used.}

\item{output.dir}{The directory where Slingshot analysis results and plots will be saved.}
}
\value{
None. This function generates and saves Slingshot analysis results and trajectory plots.
}
\description{
Function \code{run_slingshot} performs Slingshot analysis on single-cell RNA-seq data to infer cell trajectories and visualize lineage relationships.
}
\details{
Slingshot is used to infer cell trajectories and lineage relationships from single-cell RNA-seq data. This function performs the following steps:

1. Constructs a Slingshot object using PCA embeddings, cell types, start clusters, and end clusters.
2. Computes and plots the trajectory curves.
3. Computes and plots pseudotime values along the trajectory.

For more details on Slingshot, refer to the package documentation and vignettes. https://bioconductor.org/packages/devel/bioc/vignettes/slingshot/inst/doc/vignette.html and https://bustools.github.io/BUS_notebooks_R/slingshot.html
}
\author{
Zhenyi Wang <wangzy17@tsinghua.org.cn> and Yuxin Miao <miaoyx21@mails.tsinghua.edu.cn>
}
