% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_Step4_CNV_analysis.R
\name{sc_CNV}
\alias{sc_CNV}
\title{Run copykat for scRNA-seq data}
\usage{
sc_CNV(
  sc_object = NULL,
  save_path = NULL,
  assay = "RNA",
  LOW.DR = 0.05,
  UP.DR = 0.1,
  win.size = 25,
  distance = "euclidean",
  genome = "hg20",
  n.cores = 1,
  species = "hsa"
)
}
\arguments{
\item{sc_object}{The Seurat object}

\item{save_path}{The path to save results}

\item{assay}{The assay to use}

\item{...}{Parameters of `copykat::copykat`}
}
\description{
Run copykat for scRNA-seq data
}
