% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_st_10x_visium_pipeline.R,
%   R/st_10x_visium_mainPipeline.R
\name{st_10x_visium_pipeline}
\alias{st_10x_visium_pipeline}
\title{The pipeline for analyzing 10x Visium data.}
\usage{
st_10x_visium_pipeline(
  input.data.dir,
  output.dir,
  sampleName = "Hema_ST",
  Step2_QC = TRUE,
  min.gene = 200,
  min.nUMI = 500,
  max.gene = Inf,
  max.nUMI = Inf,
  min.spot = 0,
  bool.remove.mito = FALSE,
  species = "mouse",
  Step3_Clustering = TRUE,
  normalization.method = "SCTransform",
  npcs = 50,
  pcs.used = 1:10,
  resolution = 0.8,
  Step4_Find_DEGs = TRUE,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  Step5_SVFs = TRUE,
  selection.method = "moransi",
  n.top.show = 10,
  n.col.show = 5,
  Step6_Interaction = TRUE,
  commot.signaling_type = "Secreted Signaling",
  commot.database = "CellChat",
  commot.min_cell_pct = 0.05,
  commot.dis_thr = 500,
  commot.n_permutations = 100,
  Step7_CNV = TRUE,
  copykat.genome = NULL,
  copykat.LOW.DR = 0.05,
  copykat.UP.DR = 0.1,
  copykat.win.size = 25,
  copykat.distance = "euclidean",
  copykat.n.cores = 1,
  Step8_Deconvolution = TRUE,
  cell2loc.sc.h5ad.dir = NULL,
  cell2loc.sc.max.epoch = 1000,
  cell2loc.st.max.epoch = 10000,
  cell2loc.use.gpu = FALSE,
  Step9_Cellcycle = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  verbose = FALSE,
  genReport = TRUE,
  pythonPath = NULL
)

st_10x_visium_pipeline(
  input.data.dir,
  output.dir,
  sampleName = "Hema_ST",
  Step2_QC = TRUE,
  min.gene = 200,
  min.nUMI = 500,
  max.gene = Inf,
  max.nUMI = Inf,
  min.spot = 0,
  bool.remove.mito = FALSE,
  species = "mouse",
  Step3_Clustering = TRUE,
  normalization.method = "SCTransform",
  npcs = 50,
  pcs.used = 1:10,
  resolution = 0.8,
  Step4_Find_DEGs = TRUE,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  Step5_SVFs = TRUE,
  selection.method = "moransi",
  n.top.show = 10,
  n.col.show = 5,
  Step6_Interaction = TRUE,
  commot.signaling_type = "Secreted Signaling",
  commot.database = "CellChat",
  commot.min_cell_pct = 0.05,
  commot.dis_thr = 500,
  commot.n_permutations = 100,
  Step7_CNV = TRUE,
  copykat.genome = NULL,
  copykat.LOW.DR = 0.05,
  copykat.UP.DR = 0.1,
  copykat.win.size = 25,
  copykat.distance = "euclidean",
  copykat.n.cores = 1,
  Step8_Deconvolution = TRUE,
  cell2loc.sc.h5ad.dir = NULL,
  cell2loc.sc.max.epoch = 1000,
  cell2loc.st.max.epoch = 10000,
  cell2loc.use.gpu = FALSE,
  Step9_Cellcycle = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  verbose = FALSE,
  genReport = TRUE,
  pythonPath = NULL
)
}
\arguments{
\item{output.dir}{A character of path to store the results and figures}

\item{sampleName}{A character of the name of the sample}

\item{min.gene}{An integer representing the minimum number of genes detected in a spot}

\item{min.nUMI}{An integer representing the minimum number of nUMI detected in a spot}

\item{max.gene}{An integer representing the maximum number of genes detected in a spot}

\item{max.nUMI}{An integer representing the maximum number of nUMI detected in a spot}

\item{min.spot}{An integer representing the minimum number of spots expressing each gene}

\item{bool.remove.mito}{A bool value indicating whether removing mitochondrial gene}

\item{species}{A character representing the species of sample, `human` or `mouse`}

\item{normalization.method}{'SCTransform' or other choice of `normalization.method` in `NormalizaData`}

\item{resolution}{An float parameter of `FindClusters` of `Seurat`}

\item{s.features}{A gene vector used to calculate the S phase score. If NULL,
the default gene set will be used}

\item{g2m.features}{A gene vector used to calculate the G2/M phase score. If NULL,
the default gene set will be used}

\item{verbose}{verbose as `Seurat`}

\item{genReport}{A bool value indicating whether generating the report}

\item{input.data.dirs}{A character of data path, where there are filtered_feature_bc_matrix.h5
and spatial folder}

\item{n.dim.used}{An integer representing the number of dimension used for finding neighbors}

\item{bool.regress.cycling}{A bool value indicating whether regressing cycling}

\item{bool.regress.cycling.standard}{A bool value indicating the method using in
regressing cycling. See more at `https://satijalab.org/seurat/articles/cell_cycle_vignette.html`}
}
\value{
Return the user with a well-formatted HTML analysis report.
Additionally, meticulously organize and provide the analyzed results and
publication-quality vector images for each step of the analysis workflow.

Return the user with a well-formatted HTML analysis report.
Additionally, meticulously organize and provide the analyzed results and
publication-quality vector images for each step of the analysis workflow.
}
\description{
Function \code{st_10x_visium_pipeline} encompasses the complete downstream
analysis pipeline for 10x Visium data within HemascopeR.

Function \code{st_10x_visium_pipeline} encompasses the complete downstream
analysis pipeline for 10x Visium data within HemascopeR.
}
\details{
This workflow encompasses data quality control (spot QC and gene QC),
normalization, PCA dimensionality reduction, clustering and visualization

This workflow encompasses data quality control (spot QC and gene QC),
normalization, PCA dimensionality reduction, clustering and visualization
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
