% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step9_Cellcycle.R
\name{st_Cell_cycle}
\alias{st_Cell_cycle}
\title{Calculate cell cycle scores}
\usage{
st_Cell_cycle(
  st_obj,
  save_path = ".",
  s.features = NULL,
  g2m.features = NULL,
  species = "human",
  FeatureColors.bi = colorRampPalette(colors = rev(x = brewer.pal(n = 11, name =
    "RdYlBu")))
)
}
\arguments{
\item{st_obj}{The Seurat object}

\item{save_path}{The path to save results}

\item{s.features}{A vector of features associated with S phase}

\item{g2m.features}{A vector of features associated with G2M phase}

\item{species}{A character, `human` or `mouse`}

\item{FeatureColors.bi}{A function that interpolates a set of given colors to create new color palettes and color ramps}
}
\description{
Calculate cell cycle scores
}
