% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step3_Clustering.R
\name{st_Clustering}
\alias{st_Clustering}
\title{Normalization, PCA, clustering and visualization}
\usage{
st_Clustering(
  st_obj,
  output.dir = ".",
  normalization.method = "SCTransform",
  npcs = 50,
  pcs.used = 1:10,
  resolution = 0.8,
  verbose = FALSE
)
}
\arguments{
\item{st_obj}{The SeuratObject of 10X visium data}

\item{output.dir}{A character of path to store the results and figures}

\item{normalization.method}{`SCTransform` or other choice of `normalization.method` in `NormalizaData`}

\item{npcs}{The number of dimensions for analysis}

\item{pcs.used}{The principal components (PCs) to use for analysis}

\item{resolution}{An float parameter of `FindClusters` of `Seurat`}

\item{verbose}{verbose as `Seurat`}
}
\description{
Normalization, PCA, clustering and visualization
}
\details{
This function performs normalization, PCA, clustering and visualization on the
SeuratObject of 10X visium data and returns the processed SeuratObject
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
