% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step7_CNV_analysis.R
\name{st_CNV}
\alias{st_CNV}
\title{Run copykat}
\usage{
st_CNV(
  st_obj,
  save_path,
  assay = "Spatial",
  LOW.DR = 0.05,
  UP.DR = 0.1,
  win.size = 25,
  distance = "euclidean",
  genome = "hg20",
  n.cores = 1,
  species = "human"
)
}
\arguments{
\item{st_obj}{The Seurat object}

\item{save_path}{The path to save results}

\item{assay}{The assay to use}

\item{...}{Parameters of `copykat::copykat`}
}
\description{
Run copykat
}
