% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step8_Deconvolution.R
\name{st_Deconvolution}
\alias{st_Deconvolution}
\title{The pipeline for integrated analysis of 10x visium data and scRNA-seq data}
\usage{
st_Deconvolution(
  st.data.dir,
  sc.h5ad.dir = NULL,
  library_id = "Hema_ST",
  st_obj = NULL,
  save_path = NULL,
  sc.labels.key = "seurat_clusters",
  species = "mouse",
  sc.max.epoch = 1000,
  st.max.epoch = 10000,
  use.gpu = TRUE,
  condaenv = "r-reticulate",
  use.Dataset = "LymphNode"
)
}
\arguments{
\item{st.data.dir}{A character of data path, where there are filtered_feature_bc_matrix.h5
and spatial folder}

\item{sc.h5ad.dir}{A character of path of h5ad file of scRNA-seq reference data}

\item{sc.labels.key}{A character indicating the key of scRNA-seq data labels
used during running cell2location}

\item{species}{A character, `human` or `mouse`}

\item{sc.max.epoch}{A integer representing the maximum epochs of training scRNA-seq data}

\item{st.max.epoch}{A integer representing the maximum epochs of training ST data}

\item{use.gpu}{A bool value indicating whether to use GPU}

\item{condaenv}{A character of the name of conda environment used here}

\item{use.Dataset}{'HematoMap' or 'LymphNode'}

\item{save.path}{A character of path to save data}
}
\description{
The pipeline for integrated analysis of 10x visium data and scRNA-seq data
}
