% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step4_Find_DEGs.R
\name{st_Find_DEGs}
\alias{st_Find_DEGs}
\title{Find DEGs of clusters}
\usage{
st_Find_DEGs(
  st_obj,
  output.dir = ".",
  ident.label = "seurat_clusters",
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  verbose = FALSE
)
}
\arguments{
\item{st_obj}{The SeuratObject of 10X visium data}

\item{output.dir}{A character of path to store the results and figures}

\item{ident.label}{A character of the label set to `active.ident`}

\item{only.pos}{A bool value to indicate whether only return positive markers}

\item{min.pct}{The parameter of `FindAllMarkers`}

\item{logfc.threshold}{The parameter of `FindAllMarkers`}

\item{test.use}{The test used in `FindAllMarkers`}

\item{verbose}{verbose as `Seurat`}
}
\description{
Find DEGs of clusters
}
\details{
This function finds and visualizes markers for each of the identity classes and
returns the marker dataframe.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
