% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step6_Interaction.R
\name{st_Interaction}
\alias{st_Interaction}
\title{Run spatial interaction}
\usage{
st_Interaction(
  st_data_path,
  metadata_path,
  library_id = "Hema_ST",
  label_key = "seurat_clusters",
  save_path = ".",
  species = "mouse",
  signaling_type = "Secreted Signaling",
  database = "CellChat",
  min_cell_pct = 0.05,
  dis_thr = 500,
  n_permutations = 100,
  pythonPath = NULL
)
}
\arguments{
\item{st_data_path}{The path of original ST data}

\item{metadata_path}{The path of metadata.csv}

\item{label_key}{The name of the label to determine the clusters}

\item{save_path}{The path to save output files}

\item{species}{The species of the sample, 'human' or 'mouse'}

\item{signaling_type}{The parameter of `Commot` to determine the type of interaction.
Choose from 'Secreted Signaling', 'Cell-Cell Contact', and 'ECM-Receptor' for CellChatDB or
'Secreted Signaling' and 'Cell-Cell Contact' for CellPhoneDB_v4.0. If None, all pairs in the database are returned.}

\item{database}{The parameter of `Commot` to determine the database used. 'CellChat' or 'CellPhoneDB_v4.0'}

\item{min_cell_pct}{The parameter of `Commot`. The minimum expression percentage required for LR pairs to be kept}

\item{dis_thr}{The parameter of `Commot`. The threshold of spatial distance of signaling}

\item{n_permutations}{The parameter of `Commot`. Number of label permutations for computing the p-value}

\item{pythonPath}{The path to the Python environment to use for the analysis.}
}
\description{
Run spatial interaction
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
