% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step1_Loading_Data.R
\name{st_Loading_Data}
\alias{st_Loading_Data}
\title{Loading data}
\usage{
st_Loading_Data(
  input.data.dir,
  output.dir = ".",
  sampleName = "Hema_ST",
  rds.file = FALSE,
  filename = "filtered_feature_bc_matrix.h5",
  assay = "Spatial",
  slice = "slice1",
  filter.matrix = TRUE,
  to.upper = FALSE
)
}
\arguments{
\item{output.dir}{A character of path to store the h5ad file}

\item{sampleName}{A character of the name of the sample}

\item{rds.file}{description}

\item{input.data.dirs}{A character of data path, where there are filtered_feature_bc_matrix.h5
and spatial folder}
}
\description{
Loading data
}
\details{
This function loads data and generates corresponding h5ad file
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
