% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_Step10_NicheAnalysis.R
\name{st_NicheAnalysis}
\alias{st_NicheAnalysis}
\title{Niche analysis}
\usage{
st_NicheAnalysis(
  st_obj,
  features,
  save_path = ".",
  coexistence.method = "correlation",
  kmeans.n = 4,
  st_data_path = NULL,
  slice = "slice1",
  species = "mouse",
  pythonPath = NULL
)
}
\arguments{
\item{st_obj}{SeuratObject}

\item{features}{A vector of features}

\item{save_path}{Save path}

\item{kmeans.n}{The number of clusters to cluster}

\item{st_data_path}{A path containing `spatial` file and `filtered_feature_bc_matrix.h5` file}

\item{slice}{The slice to use}

\item{species}{The species this sample belongs to}

\item{pythonPath}{The path to the Python environment to use for the analysis.}
}
\description{
Niche analysis
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
