% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_stereo_pipeline.R
\name{st_stereo_pipeline}
\alias{st_stereo_pipeline}
\title{The pipeline for analyzing stereo-seq data.}
\usage{
st_stereo_pipeline(
  input.data.dir,
  output.dir,
  sampleName = "Hema_stereo",
  data_type = "gem",
  sep = "\\t",
  bin_type = "bins",
  bin_size = 100,
  spot_diameter = 80,
  is_sparse = TRUE,
  gene_list = NULL,
  region = NULL,
  assay = "Spatial",
  Step2_QC = TRUE,
  min.gene = 20,
  min.nUMI = 50,
  max.gene = Inf,
  max.nUMI = Inf,
  min.spot = 0,
  bool.remove.mito = FALSE,
  species = "mouse",
  Step3_Clustering = TRUE,
  normalization.method = "SCTransform",
  npcs = 50,
  pcs.used = 1:10,
  resolution = 0.1,
  max.n.cluster = 30,
  Step4_Find_DEGs = TRUE,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  Step5_SVFs = TRUE,
  selection.method = "moransi",
  n.top.show = 10,
  n.col.show = 5,
  Step6_Interaction = TRUE,
  h5ad_path = NULL,
  counts_path = NULL,
  coordinates_path = NULL,
  coordinates_index_col = 0,
  counts_transpose = TRUE,
  commot.signaling_type = "Secreted Signaling",
  commot.database = "CellChat",
  commot.min_cell_pct = 0.05,
  commot.dis_thr = 500,
  commot.n_permutations = 100,
  Step7_Cellcycle = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  verbose = FALSE,
  pythonPath = NULL
)
}
\arguments{
\item{output.dir}{A character of path to store the results and figures}

\item{data_type}{'gem', 'gef' or 'h5ad'}

\item{sep}{Separator string for 'gem' format}

\item{bin_type}{The type of bin, if file format is stereo-seq file. 'bins' or 'ell_bins'.}

\item{bin_size}{the size of bin to merge. The parameter only takes effect when the value of `bin_type` is 'bins'.}

\item{is_sparse}{The matrix is sparse matrix if is_sparse is True else np.ndarray}

\item{gene_list}{Restrict to this gene list}

\item{region}{Restrict to this region, [minX, maxX, minY, maxY]}

\item{min.gene}{An integer representing the minimum number of genes detected in a spot}

\item{min.nUMI}{An integer representing the minimum number of nUMI detected in a spot}

\item{max.gene}{An integer representing the maximum number of genes detected in a spot}

\item{max.nUMI}{An integer representing the maximum number of nUMI detected in a spot}

\item{min.spot}{An integer representing the minimum number of spots expressing each gene}

\item{bool.remove.mito}{A bool value indicating whether removing mitochondrial gene}

\item{species}{A character representing the species of sample, `human` or `mouse`}

\item{Step3_Clustering}{A bool value indicating whether performing clustering.}

\item{normalization.method}{'SCTransform' or other choice of `normalization.method` in `NormalizaData`}

\item{resolution}{An float parameter of `FindClusters` of `Seurat`}

\item{Step4_Find_DEGs}{A bool value indicating whether finding DEGs.}

\item{only.pos}{A bool value to indicate whether only return positive markers.}

\item{min.pct}{The parameter of `FindAllMarkers`.}

\item{logfc.threshold}{The parameter of `FindAllMarkers`.}

\item{test.use}{The test used in `FindAllMarkers`.}

\item{Step5_SVFs}{A bool value indicating whether identifying spatially variable genes.}

\item{selection.method}{Method for selecting spatially variable features,
`markvariogram` or `moransi`}

\item{n.top.show}{The number of top genes shown in the figure}

\item{Step6_Interaction}{A bool value indicating whether performing spatial interaction analysis.}

\item{h5ad_path}{The path of MERFISH data in h5ad format}

\item{counts_path}{The path of counts data in csv format}

\item{coordinates_path}{The path of coordinates data in xlsx format}

\item{coordinates_index_col}{The column index of coordinates file}

\item{counts_transpose}{Whether to transpose the counts matrix. If rows stand for genes, it will be `FALSE`.}

\item{commot.signaling_type}{The parameter of `Commot` to determine the type of interaction.
Choose from 'Secreted Signaling', 'Cell-Cell Contact', and 'ECM-Receptor' for CellChatDB or
'Secreted Signaling' and 'Cell-Cell Contact' for CellPhoneDB_v4.0. If None, all pairs in the database are returned.}

\item{commot.database}{The parameter of `Commot` to determine the database used. 'CellChat' or 'CellPhoneDB_v4.0'}

\item{commot.min_cell_pct}{The parameter of `Commot`. The minimum expression percentage required for LR pairs to be kept}

\item{commot.dis_thr}{The parameter of `Commot`. The threshold of spatial distance of signaling}

\item{commot.n_permutations}{The parameter of `Commot`. Number of label permutations for computing the p-value}

\item{verbose}{verbose as `Seurat`}

\item{pythonPath}{The path to the Python environment to use for the analysis.}

\item{input.data.dirs}{A character of data path.}

\item{n.dim.used}{An integer representing the number of dimension used for finding neighbors}

\item{n.col}{The number of columns in the output figure}

\item{@param}{Step2_QC A bool value indicating whether performing QC.}
}
\description{
Function \code{st_stereo_pipeline} encompasses the complete downstream
analysis pipeline for four kinds of MERFISH data within HemascopeR.
}
\details{
The st_MERFISH_pipeline function encapsulates the complete downstream analysis pipeline
for MERFISH data within HemascopeR. It takes in various parameters to customize the analysis
workflow and provided analyzed results and publication-quality vector images for each step of the analysis workflow.

The function performs the following steps:

1. Loading data.

2. QC (Quality Control) of genes and spots.

3. Normalization, PCA, and clustering.

4. Finding differentially expressed genes in each cluster.

5. Identifying spatially variable features.

6. Performing spatial interaction analysis using Commot.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
