% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereo_Step1_Loading_Data.R
\name{stereo_Loading_Data}
\alias{stereo_Loading_Data}
\title{Loading stereo-seq data}
\usage{
stereo_Loading_Data(
  input.data.dir,
  output.dir = ".",
  sampleName = "Hema_ST",
  assay = "Spatial",
  data_type = "gem",
  sep = "\\t",
  bin_type = "bins",
  bin_size = 100,
  spot_diameter = 80,
  is_sparse = TRUE,
  gene_list = NULL,
  region = NULL,
  pythonPath = NULL
)
}
\arguments{
\item{output.dir}{A character of path to store the h5ad file}

\item{data_type}{'gem', 'gef' or 'h5ad'}

\item{sep}{Separator string for 'gem' format}

\item{bin_type}{The type of bin, if file format is stereo-seq file. 'bins' or 'ell_bins'.}

\item{bin_size}{the size of bin to merge. The parameter only takes effect when the value of `bin_type` is 'bins'.}

\item{is_sparse}{The matrix is sparse matrix if is_sparse is True else np.ndarray}

\item{gene_list}{Restrict to this gene list}

\item{region}{Restrict to this region, [minX, maxX, minY, maxY]}

\item{pythonPath}{The path to the Python environment to use for the analysis}

\item{input.data.dirs}{A character of data path}
}
\description{
Loading stereo-seq data
}
\details{
This function loads stereo-seq data in different formats
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
