% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_10x_chromium_mainPipeline.R
\name{scRNASeq_10x_pipeline}
\alias{scRNASeq_10x_pipeline}
\title{Comprehensive Pipeline for Analyzing 10x Single-Cell RNA-Seq Data}
\usage{
scRNASeq_10x_pipeline(
  input.data.dirs = NULL,
  project.names = NULL,
  output.dir = NULL,
  pythonPath = python.path.sc(),
  databasePath = NULL,
  gene.column = 2,
  min.cells = 10,
  min.feature = 200,
  mt.pattern = "^mt-",
  nFeature_RNA.limit = 200,
  percent.mt.limit = 20,
  scale.factor = 10000,
  nfeatures = 3000,
  ndims = 50,
  vars.to.regress = NULL,
  PCs = 1:35,
  resolution = 0.4,
  n.neighbors = 50,
  doublet.percentage = 0.04,
  doublerFinderwraper.PCs = 1:20,
  doublerFinderwraper.pN = 0.25,
  doublerFinderwraper.pK = 0.1,
  phate.knn = 50,
  phate.npca = 20,
  phate.t = 10,
  phate.ndim = 2,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  marker.genes = NULL,
  ViolinPlot.cellTypeOrders = NULL,
  ViolinPlot.cellTypeColors = NULL,
  Org = NULL,
  lineage.genelist = NULL,
  lineage.names = NULL,
  groups_colors = NULL,
  your_openai_API_key = "",
  tissuename = "",
  gptmodel = "gpt-3.5",
  slingshot.start.clus = NULL,
  slingshot.end.clus = NULL,
  slingshot.colors = NULL,
  loom.files.path = NULL,
  cellcycleCutoff = NULL,
  sorting = FALSE,
  ncores = 1,
  Whether_load_previous_results = FALSE,
  Step1_Input_Data = TRUE,
  Step1_Input_Data.type = NULL,
  Step2_Quality_Control = TRUE,
  Step2_Quality_Control.RemoveBatches = TRUE,
  Step2_Quality_Control.RemoveDoublets = TRUE,
  Step3_Clustering = TRUE,
  Step4_Identify_Cell_Types = TRUE,
  Step4_Use_Which_Labels = NULL,
  Step4_Cluster_Labels = NULL,
  Step4_Changed_Labels = NULL,
  Step4_run_sc_CNV = TRUE,
  Step5_Visualization = TRUE,
  Step6_Find_DEGs = TRUE,
  Step7_Assign_Cell_Cycle = TRUE,
  Step8_Calculate_Heterogeneity = TRUE,
  Step9_Violin_Plot_for_Marker_Genes = TRUE,
  Step10_Calculate_Lineage_Scores = TRUE,
  Step11_GSVA = TRUE,
  Step11_GSVA.identify.cellType.features = TRUE,
  Step11_GSVA.identify.diff.features = FALSE,
  Step11_GSVA.comparison.design = NULL,
  Step12_Construct_Trajectories = TRUE,
  Step12_Construct_Trajectories.clusters = NULL,
  Step12_Construct_Trajectories.monocle = TRUE,
  Step12_Construct_Trajectories.slingshot = TRUE,
  Step12_Construct_Trajectories.scVelo = TRUE,
  Step13_TF_Analysis = TRUE,
  Step13_TF_Analysis.cellTypes_colors = NULL,
  Step13_TF_Analysis.groups_colors = NULL,
  Step14_Cell_Cell_Interaction = TRUE,
  Step15_Generate_the_Report = TRUE
)
}
\arguments{
\item{input.data.dirs}{A character vector containing directory paths where 10x single-cell RNA-Seq data is stored. Each element corresponds to a path for a dataset.}

\item{project.names}{A character vector specifying project names for each dataset in input.data.dirs. Its length should match the number of datasets in input.data.dirs, providing project names for each dataset. Please note that the 'project.names' parameter should be the same as the 'id' parameter in cellranger count.}

\item{output.dir}{The directory where analysis results and reports will be stored.}

\item{pythonPath}{The path to the Python environment to use for the analysis.}

\item{databasePath}{The path to the database required for the analysis.}

\item{gene.column}{The column number in the 10x data where gene information is stored. (Default: 2)}

\item{min.cells}{The minimum number of cells for a gene to be considered. (Default: 10)}

\item{min.feature}{The minimum number of features (genes) for a cell to be considered. (Default: 200)}

\item{mt.pattern}{The regular expression pattern to identify mitochondrial genes. (Default: '^mt-')}

\item{nFeature_RNA.limit}{The minimum number of features (genes) in the cell. (Default: 200)}

\item{percent.mt.limit}{The maximum percentage of mitochondrial genes allowed. (Default: 20)}

\item{scale.factor}{The factor for data scaling. (Default: 10000)}

\item{nfeatures}{The number of highly variable features (genes) to consider. (Default: 3000)}

\item{ndims}{The number of dimensions for downstream analysis. (Default: 50)}

\item{vars.to.regress}{Variables to regress during data normalization. (Default: NULL)}

\item{PCs}{The principal components (PCs) to use for downstream analysis. (Default: 1:35)}

\item{resolution}{The resolution parameter for Louvain clustering. (Default: 0.4)}

\item{n.neighbors}{The number of neighbors for downstream analysis. (Default: 50)}

\item{doublet.percentage}{The percentage used for doublet removal. (Default: 0.04)}

\item{doublerFinderwraper.PCs}{The PCs used in the doublet removal process. (Default: 1:20)}

\item{doublerFinderwraper.pN}{The pN parameter for doublet removal. (Default: 0.25)}

\item{doublerFinderwraper.pK}{The pK parameter for doublet removal. (Default: 0.1)}

\item{phate.knn}{The number of nearest neighbors for PhateR analysis. (Default: 50)}

\item{phate.npca}{The number of principal components for PhateR analysis. (Default: 20)}

\item{phate.t}{The t parameter for PhateR analysis. (Default: 10)}

\item{phate.ndim}{The number of dimensions for PhateR analysis. (Default: 2)}

\item{min.pct}{The minimum percentage for a gene to be considered in differential gene detection. (Default: 0.25)}

\item{logfc.threshold}{The log-fold change threshold for differential gene detection. (Default: 0.25)}

\item{marker.genes}{A character vector specifying marker genes for violin plot visualization. (Default: NULL)}

\item{ViolinPlot.cellTypeOrders}{The order of cell types for violin plot visualization. (Default: NULL)}

\item{ViolinPlot.cellTypeColors}{The colors for cell types in violin plot visualization. (Default: NULL)}

\item{Org}{The organism for analysis (e.g., 'mmu' for mouse, 'hsa' for human). (Default: NULL)}

\item{lineage.genelist}{A list of gene names for defining cell lineages. (Default: NULL)}

\item{lineage.names}{A character vector specifying names for each lineage in lineage.genelist. (Default: NULL)}

\item{groups_colors}{A named list specifying colors for groups. (Default: NULL)}

\item{your_openai_API_key}{The openai_key parameter for gptcelltype. The OpenAI key obtained from https://platform.openai.com/account/api-keys The default is NA, which will resulting outputing the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.}

\item{tissuename}{The tissuename parameter for gptcelltype. Optional input of tissue name. Default = ''.}

\item{gptmodel}{The model parameter for gptcelltype. A valid GPT-4 or GPT-3.5 model name list on https://platform.openai.com/docs/models. Default is 'gpt-3.5'.}

\item{slingshot.start.clus}{The starting cluster for slingshot analysis. (Default: NULL)}

\item{slingshot.end.clus}{The ending cluster for slingshot analysis. (Default: NULL)}

\item{slingshot.colors}{The colors for data points in slingshot visualization. (Default: NULL)}

\item{loom.files.path}{The path of loom files for scVelo analysis. (Default: NULL)}

\item{cellcycleCutoff}{The cell cycle cutoff for cell cycle assignment (Default: NULL).}

\item{sorting}{Boolean indicating whether the cells were sorted. (Default: FALSE)}

\item{ncores}{The number of CPU cores to use for parallel processing. (Default: 1)}

\item{Whether_load_previous_results}{Boolean indicating whether to load previous analysis results. (Default: FALSE)}

\item{Step1_Input_Data}{Boolean indicating whether to perform Step 1: Input Data. (Default: TRUE)}

\item{Step1_Input_Data.type}{The type of the input data, e.g. '10x' for cellranger-count software output, 'Seurat' for Seurat object in Rds format, and 'Matrix' for gene expression matrix in Rds format. (Default: NULL)}

\item{Step2_Quality_Control}{Boolean indicating whether to perform Step 2: Quality Control. (Default: TRUE)}

\item{Step2_Quality_Control.RemoveBatches}{Boolean indicating whether to remove batches during quality control. (Default: TRUE)}

\item{Step2_Quality_Control.RemoveDoublets}{Boolean indicating whether to remove doublets during quality control. (Default: TRUE)}

\item{Step3_Clustering}{Boolean indicating whether to perform Step 3: Clustering. (Default: TRUE)}

\item{Step4_Identify_Cell_Types}{Boolean indicating whether to perform Step 4: Identify Cell Types. (Default: TRUE)}

\item{Step4_Use_Which_Labels}{The character specifies which labels to use for cell type identification in Step 4, including 'clustering', 'abcCellmap.1', 'abcCellmap.2', 'abcCellmap.3', 'abcCellmap.4', 'HematoMap', and 'changeLabels'. (Default: 'clustering')}

\item{Step4_Cluster_Labels}{Characters. If Step4_Use_Which_Labels parameter was set to 'changeLabels', please set Step4_Cluster_Labels parameter as the characters indicating the clustering labels. (Default: NULL)}

\item{Step4_Changed_Labels}{Characters. If Step4_Use_Which_Labels parameter was set to 'changeLabels', please set Step4_Changed_Labels parameter as the characters indicating the changed labels. Note that the length of the character vectors of Step4_Cluster_Labels should be equal to that of Step4_Changed_Labels. (Default: NULL)}

\item{Step4_run_sc_CNV}{Boolean indicating whether to run single-cell copy number variation analysis in Step 4. (Default: FALSE)}

\item{Step5_Visualization}{Boolean indicating whether to perform Step 5: Visualization. (Default: TRUE)}

\item{Step6_Find_DEGs}{Boolean indicating whether to perform Step 6: Find Differentially Expressed Genes (DEGs). (Default: TRUE)}

\item{Step7_Assign_Cell_Cycle}{Boolean indicating whether to perform Step 7: Assign Cell Cycle. (Default: TRUE)}

\item{Step8_Calculate_Heterogeneity}{Boolean indicating whether to perform Step 8: Calculate Heterogeneity. (Default: TRUE)}

\item{Step9_Violin_Plot_for_Marker_Genes}{Boolean indicating whether to perform Step 9: Violin Plot for Marker Genes. (Default: TRUE)}

\item{Step10_Calculate_Lineage_Scores}{Boolean indicating whether to perform Step 10: Calculate Lineage Scores. (Default: TRUE)}

\item{Step11_GSVA}{Boolean indicating whether to perform Step 11: Gene Set Variation Analysis (GSVA). (Default: TRUE)}

\item{Step11_GSVA.identify.cellType.features}{Boolean indicating whether to identify cell type-specific features in GSVA. (Default: TRUE)}

\item{Step11_GSVA.identify.diff.features}{Boolean indicating whether to identify differentially expressed features in GSVA. (Default: TRUE)}

\item{Step11_GSVA.comparison.design}{A list specifying the comparison design for GSVA. (Default: NULL)}

\item{Step12_Construct_Trajectories}{Boolean indicating whether to perform Step 12: Construct Trajectories. (Default: TRUE)}

\item{Step12_Construct_Trajectories.clusters}{A vector indicating the clusters used for trajectory construction in Step 12. (Default: NULL)}

\item{Step12_Construct_Trajectories.monocle}{Boolean indicating whether to use Monocle for trajectory construction in Step 12. (Default: FALSE)}

\item{Step12_Construct_Trajectories.slingshot}{Boolean indicating whether to use slingshot for trajectory construction in Step 12. (Default: FALSE)}

\item{Step12_Construct_Trajectories.scVelo}{Boolean indicating whether to use scVelo for trajectory construction in Step 12. (Default: FALSE)}

\item{Step13_TF_Analysis}{Boolean indicating whether to perform Step 13: Transcription Factor (TF) Analysis. (Default: TRUE)}

\item{Step13_TF_Analysis.cellTypes_colors}{A character vector indicating cell type-specific colors in TF analysis in Step 13. (Default: NULL)}

\item{Step13_TF_Analysis.groups_colors}{A character vector indicating group-specific colors in TF analysis in Step 13. (Default: NULL)}

\item{Step14_Cell_Cell_Interaction}{Boolean indicating whether to perform Step 14: Cell-Cell Interaction. (Default: TRUE)}

\item{Step15_Generate_the_Report}{Boolean indicating whether to perform Step 15: Generate the Report. (Default: TRUE)}
}
\value{
Returns the user with a well-formatted HTML analysis report and organizes the analyzed.
results and publication-quality vector images for each step of the analysis workflow.
}
\description{
The function \code{scRNASeq_10x_pipeline} provides a comprehensive analysis
pipeline for 10x single-cell RNA-Seq data within the HemascopeR package.
}
\details{
This pipeline offers a comprehensive workflow for the analysis of 10x single-cell
RNA-Seq data. It includes the following steps:

1. Input Data: Loading and preprocessing of single-cell RNA-Seq data from specified
directories.

2. Quality Control: Performing data quality control, filtering low-quality cells
and features, and calculating metrics like the percentage of mitochondrial genes.

3. Clustering: Identifying cell clusters and visualizing them using t-SNE and UMAP
dimensionality reduction.

4. Automated Cell Type Identification: Automatically assigning cell types to clusters
based on marker genes and mapping single-cell data to an internal reference atlas of
hematopoietic cells.

5. Visualization: Run phateR for additional dimensionality reduction and visualization.

6. Differential Gene Detection: Identifying differentially expressed genes (DEGs) between
cell clusters.

7. Cell Cycle Assignment: Assigning cells to specific cell cycle phases.

8. Assessment of Heterogeneity: Assessing heterogeneity within cell populations.

9. Creating violin plots to explore marker genes and cluster characteristics.

10. Prediction of Lineage Scores: Predicting lineage scores for individual cells.

11. Gene Set Variation Analysis (GSVA): Calculating gene set variation scores for cell types.

12. Cell Differentiation Trajectory Prediction: Predicting cell differentiation trajectories
using Monocle2, Slingshot and scVelo.

13. Transcription Factor (TF) Analysis: Analyzing transcription factors and their activity.

14. Cell-Cell Interactions: Analyzing cell-cell interactions and communication.

15. Generate a well-formatted HTML analysis report.

Note that, if you will use Step 5 or scVelo in Step 12, please ensure that you have installed required Python-packages in your Python environment and specify
the Python path using the 'pythonPath' parameter.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
