% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_10x_visium_mainPipeline.R
\name{st_10x_visium_pipeline}
\alias{st_10x_visium_pipeline}
\title{The pipeline for analyzing 10x Visium data.}
\usage{
st_10x_visium_pipeline(
  input.data.dir,
  output.dir,
  sampleName = "Hema_ST",
  rds.file = FALSE,
  filename = "filtered_feature_bc_matrix.h5",
  assay = "Spatial",
  slice = "slice1",
  filter.matrix = TRUE,
  to.upper = FALSE,
  Step2_QC = TRUE,
  min.gene = 200,
  min.nUMI = 500,
  max.gene = Inf,
  max.nUMI = Inf,
  min.spot = 0,
  bool.remove.mito = FALSE,
  species = "mouse",
  Step3_Clustering = TRUE,
  normalization.method = "SCTransform",
  npcs = 50,
  pcs.used = 1:10,
  resolution = 0.8,
  Step4_Find_DEGs = TRUE,
  only.pos = TRUE,
  min.pct = 0.25,
  logfc.threshold = 0.25,
  test.use = "wilcox",
  Step5_SVFs = TRUE,
  selection.method = "moransi",
  n.top.show = 10,
  n.col.show = 5,
  Step6_Interaction = TRUE,
  commot.signaling_type = "Secreted Signaling",
  commot.database = "CellChat",
  commot.min_cell_pct = 0.05,
  commot.dis_thr = 500,
  commot.n_permutations = 100,
  Step7_CNV = TRUE,
  copykat.genome = NULL,
  copykat.LOW.DR = 0.05,
  copykat.UP.DR = 0.1,
  copykat.win.size = 25,
  copykat.distance = "euclidean",
  copykat.n.cores = 1,
  Step8_Deconvolution = TRUE,
  cell2loc.sc.h5ad.dir = NULL,
  cell2loc.sc.max.epoch = 1000,
  cell2loc.st.max.epoch = 10000,
  cell2loc.use.gpu = FALSE,
  cell2loc.use.Dataset = "LymphNode",
  Step9_Cellcycle = TRUE,
  s.features = NULL,
  g2m.features = NULL,
  Step10_Niche = TRUE,
  coexistence.method = "correlation",
  Niche.cluster.n = 4,
  verbose = FALSE,
  genReport = TRUE,
  pythonPath = python.path.ST()
)
}
\arguments{
\item{output.dir}{A character of path to store the results and figures.}

\item{sampleName}{A character of the name of the sample.}

\item{Step2_QC}{A bool value indicating whether performing QC.}

\item{min.gene}{An integer representing the minimum number of genes detected in a spot.}

\item{min.nUMI}{An integer representing the minimum number of nUMI detected in a spot.}

\item{max.gene}{An integer representing the maximum number of genes detected in a spot.}

\item{max.nUMI}{An integer representing the maximum number of nUMI detected in a spot.}

\item{min.spot}{An integer representing the minimum number of spots expressing each gene.}

\item{bool.remove.mito}{A bool value indicating whether removing mitochondrial gene.}

\item{species}{A character representing the species of sample, `human` or `mouse`.}

\item{Step3_Clustering}{A bool value indicating whether performing clustering.}

\item{normalization.method}{'SCTransform' or other choice of `normalization.method` in `NormalizaData`.}

\item{npcs}{The total number of PCs to compute.}

\item{pcs.used}{Dimensions of PCs to use as input.}

\item{resolution}{An float parameter of `FindClusters` of `Seurat`.}

\item{Step4_Find_DEGs}{A bool value indicating whether finding DEGs.}

\item{only.pos}{A bool value to indicate whether only return positive markers.}

\item{min.pct}{The parameter of `FindAllMarkers`.}

\item{logfc.threshold}{The parameter of `FindAllMarkers`.}

\item{test.use}{The test used in `FindAllMarkers`.}

\item{Step5_SVFs}{A bool value indicating whether identifying spatially variable genes.}

\item{selection.method}{Method for selecting spatially variable features, `markvariogram` or `moransi`.}

\item{Step6_Interaction}{A bool value indicating whether performing spatial interaction analysis.}

\item{commot.signaling_type}{The parameter of `Commot` to determine the type of interaction.
Choose from 'Secreted Signaling', 'Cell-Cell Contact', and 'ECM-Receptor' for CellChatDB or.
'Secreted Signaling' and 'Cell-Cell Contact' for CellPhoneDB_v4.0. If None, all pairs in the database are returned.}

\item{commot.database}{The parameter of `Commot` to determine the database used. 'CellChat' or 'CellPhoneDB_v4.0'.}

\item{commot.min_cell_pct}{The parameter of `Commot`. The minimum expression percentage required for LR pairs to be kept.}

\item{commot.dis_thr}{The parameter of `Commot`. The threshold of spatial distance of signaling.}

\item{commot.n_permutations}{The parameter of `Commot`. Number of label permutations for computing the p-value.}

\item{Step7_CNV}{A bool value indicating whether performing CNV analysis.}

\item{copykat.genome}{'hg20' or 'mm10'}

\item{copykat.LOW.DR}{The parameter of `copykat`. The minimal population fractions of genes for smoothing.}

\item{copykat.UP.DR}{The parameter of `copykat`. The minimal population fractions of genes for segmentation.}

\item{copykat.win.size}{The parameter of `copykat`. The minimal window sizes for segmentation.}

\item{copykat.distance}{The parameter of `copykat`. Distance methods include euclidean,
and correlation converted distance include pearson and spearman.}

\item{copykat.n.cores}{The parameter of `copykat`. The number of cores for parallel computing.}

\item{Step8_Deconvolution}{A bool value indicating whether performing deconvolution analysis.}

\item{cell2loc.sc.h5ad.dir}{A character of path of h5ad file of scRNA-seq reference data, default NULL and used default dataset.}

\item{cell2loc.sc.max.epoch}{A integer representing the maximum epochs of training scRNA-seq data.}

\item{cell2loc.st.max.epoch}{A integer representing the maximum epochs of training ST data.}

\item{cell2loc.use.Dataset}{'HematoMap' or 'LymphNode'.}

\item{Step9_Cellcycle}{A bool value indicating whether analyzing cell cycle.}

\item{s.features}{A gene vector used to calculate the S phase score. If NULL,
the default gene set will be used.}

\item{g2m.features}{A gene vector used to calculate the G2/M phase score. If NULL,
the default gene set will be used.}

\item{Step10_Niche}{A bool value indicating whether performing niche analysis.}

\item{coexistence.method}{The method to analyze the coexistence of cell types. 'correlation' or 'Wasserstein'.}

\item{Niche.cluster.n}{The number of clusters to cluster.}

\item{verbose}{verbose as `Seurat`.}

\item{genReport}{A bool value indicating whether generating the report.}

\item{pythonPath}{The path to the Python environment to use for the analysis.}

\item{input.data.dirs}{A character of data path, where there are filtered_feature_bc_matrix.h5
and spatial folder.}

\item{bool.regress.cycling}{A bool value indicating whether regressing cycling.}

\item{bool.regress.cycling.standard}{A bool value indicating the method using in
regressing cycling. See more at `https://satijalab.org/seurat/articles/cell_cycle_vignette.html`.}
}
\value{
Return the user with a well-formatted HTML analysis report.
Additionally, meticulously organize and provide the analyzed results and
publication-quality vector images for each step of the analysis workflow.
}
\description{
Function \code{st_10x_visium_pipeline} encompasses the complete downstream
analysis pipeline for 10x Visium data within HemascopeR.
}
\details{
The st_10x_visium_pipeline function encapsulates the complete downstream analysis pipeline
for 10x Visium data within HemascopeR. It takes in various parameters to customize the analysis
workflow and generates a well-formatted HTML analysis report along with meticulously organized
and provided analyzed results and publication-quality vector images for each step of the analysis workflow.

The function performs the following steps:

1. Loading data.

2. QC (Quality Control) of genes and spots.

3. Normalization, PCA, and clustering.

4. Finding differentially expressed genes in each cluster.

5. Identifying spatially variable features.

6. Performing spatial interaction analysis using Commot.

7. CNV (Copy Number Variation) analysis using copykat.

8. Deconvolution using cell2location.

9. Cell cycle analysis.

10. Niche analysis.

Finally, it saves the analyzed data and generates the report.
}
\author{
Zhenyi Wang wangzy17@tsinghua.org.cn and Yuxin Miao miaoyx21@mails.tsinghua.edu.cn
}
